Move blacklist to constants
