+++ src/analysis/retail/evoker/augmentation/constants.tsx
@@ -25,3 +25,38 @@ export const TECTONIC_LOCUS_MULTIPLIER = 0.5;
 export const VOLCANISM_ESSENCE_REDUCTION = 1;
 export const ANACHRONISM_ESSCENCE_CHANCE = 0.35;
 export const SYMBIOTIC_HEALING_INCREASE = 0.03;
+
+/** SpellIds to blacklist, ie. trinkets that doesnt add contribution */
+export const blacklist: number[] = [
+  402583, // Beacon
+  408682, // Dragonfire Bomb Dispenser
+  401324, // Pocket Anvil (Echoed Flare)
+  401306, // Pocket Anvil (Anvil Strike)
+  401422, // Vessel of Searing Shadow (Shadow Spike)
+  401428, // Vessel of Searing Shadow (Ravenous Shadowflame)
+  418774, // Mirror of Fractured Tomorrows ()
+  418588, // Mirror of Fractured Tomorrows (Sand Cleave)
+  419591, // Mirror of Fractured Tomorrows (Auto Attack)
+  418607, // Mirror of Fractured Tomorrows (Sand Bolt)
+  406251, // Roiling Shadowflame
+  406889, // Roiling Shadowflame (Self Harm)
+  400223, // Thorns of Iron
+  322109, // Touch of Death
+  124280, // Touch of Karma
+  184689, // Shield of Vengeance
+  379403, // Toxic Thorn Footwraps (Launched Thorns)
+  408791, // Ashkandur, Fall of the Brotherhood
+  378426, // Slimy Expulsion Boots boots (Corrosive Slime)
+  381006, // Acidic Hailstone Treads (Deep Chill)
+  381700, // Forgestorm (Forgestorm Ignited)
+  406764, // Shadowflame Wreathe
+  394453, // Broodkeeper's Blaze
+  370794, // Unstable Frostfire Belt (Lingering Frostspark)
+  408836, // Djaruun, Pillar of the Elder Flame
+  408815, // Djaruun, Pillar of the Elder Flame
+  381475, // Erupting Spear Fragment
+  281721, // Bile-Stained Crawg Tusks (Vile Bile)
+  214397, // Mark of Dargrul (Landslide)
+  408469, // Call to Suffering (Self Harm)
+  409632, // Breath of Eons
+];
